tests: Split tests into multiple source files
In order to have a more clearer view on the test cases, the tests
have been split into separate source files, as follows:

test_io.cpp will include the tests related to IO such as:

- R/W tests
- I/O stats
- ...

test_disk_actions.cpp will include the tests related to disk
actions such as:

- Map/Unmap operations
- Live resize
- Persistent reservations
- ...
